Remove LlmInvocationHistory dependence on spi
